Indent

+++ src/blind_rsa.c
@@ -4,7 +4,7 @@
 #include <string.h>
 
 #ifndef OPENSSL_API_COMPAT
-#define OPENSSL_API_COMPAT 10100
+#    define OPENSSL_API_COMPAT 10100
 #endif
 
 #include <openssl/bn.h>
@@ -19,7 +19,7 @@
 #include "blind_rsa.h"
 
 #ifndef OPENSSL_IS_BORINGSSL
-#define BN_bn2bin_padded(OUT, LEN, IN) (BN_bn2binpad((IN), (OUT), (LEN)) == (LEN))
+#    define BN_bn2bin_padded(OUT, LEN, IN) (BN_bn2binpad((IN), (OUT), (LEN)) == (size_t) (LEN))
 #endif
 
 #define MIN_MODULUS_BITS      2048
